feat: CI 파이프라인에 Redis 테스트 환경 구축 및 테스트 코드 추가
