+++ .github/workflows/ci.yml
@@ -10,18 +10,20 @@ jobs:
   # build(test포함) 진행
   build:
     runs-on: ubuntu-latest
-
     steps:
       - name: Github Repository 파일 불러오기
         uses: actions/checkout@v4
-
       - name: JDK 17버전 설치
         uses: actions/setup-java@v4
         with:
           distribution: temurin
           java-version: 17
-      - name: .env 파일 만들기
-        run: echo "${{ secrets.ENV }}" > .env
+      - name: Redis 설치
+        uses: supercharge/redis-github-action@1.1.0
+        with:
+          redis-version: 6
+      - name: 테스트용 .env 파일 만들기
+        run: echo "${{ secrets.ENV_TEST }}" > .env
       - name: Grant execute permission to Gradle wrapper
         run: chmod +x ./gradlew
       - name: Build and run tests
