+++ .github/workflows/deploy.yml
@@ -17,6 +17,12 @@ jobs:
         with:
           distribution: temurin
           java-version: 17
+      - name: Redis 설치
+        uses: supercharge/redis-github-action@1.1.0
+        with:
+          redis-version: 6
+      - name: 테스트용 .env 파일 만들기
+        run: echo "${{ secrets.ENV_TEST }}" > .env
       - name: 테스트 및 빌드하기
         run: |
           sudo chmod +x gradlew
