+++ src/test/java/com/pawever/server/domain/cicd/service/RedisServiceTest.java
@@ -1,60 +1,29 @@
 package com.pawever.server.domain.cicd.service;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+@SpringBootTest
+public class RedisServiceTest {
+    // Mock 객체 없이 실제 Redis 테스트
 
-class RedisServiceTest {
-    @Mock
-    private RedisTemplate<String, String> redisTemplate; // RedisTemplate을 Mock 객체로 선언
-
-    @Mock
-    private ValueOperations<String, String> valueOperations; // ValueOperations을 Mock 객체로 선언
-
-    @InjectMocks
-    private RedisService redisService; // 테스트 대상 서비스
-
-    @BeforeEach
-    public void setUp() {
-        // Mockito 초기화
-        MockitoAnnotations.initMocks(this);
-
-        // opsForValue()를 Mock 객체로 설정
-        when(redisTemplate.opsForValue()).thenReturn(valueOperations);
-    }
+    @Autowired
+    private RedisService redisService;
 
     @Test
-    public void testSaveValue() {
-        // Given: 테스트를 위한 값 설정
-        String key = "myKey";
-        String value = "myValue";
+    public void testSaveAndGetValue() {
+        // given
+        String key = "testKey";
+        String value = "testValue";
 
-        // When: saveValue 메서드 호출
+        // when
         redisService.saveValue(key, value);
+        String retrievedValue = redisService.getValue(key);
 
-        // Then: RedisTemplate의 opsForValue().set() 메서드가 호출되었는지 검증
-        verify(valueOperations).set(key, value); // set() 메서드가 호출되었는지 검증
-    }
-
-    @Test
-    public void testGetValue() {
-        // Given: RedisTemplate이 리턴할 값을 설정
-        String key = "myKey";
-        String expectedValue = "myValue";
-        when(valueOperations.get(key)).thenReturn(expectedValue); // mock 리턴 값 설정
-
-        // When: getValue 메서드 호출
-        String actualValue = redisService.getValue(key);
-
-        // Then: 값이 제대로 반환되는지 검증
-        assertEquals(expectedValue, actualValue); // 예상 값과 실제 값 비교
+        // then
+        assertThat(retrievedValue).isEqualTo(value);
+        System.out.println("Redis Test without Mock success");
     }
 }
